Extract OptionLegCard into own file.
